:MSMatchAddByDiffHlGroup: Distinct handling of DiffChange and DiffText
Previously, the plugin only considered the highlighting of the last character in a line, which is somewhat arbitrary.
Now, for DiffChange accept any change in the line, so the last character can be DiffChange (change elsewhere in the line) or DiffText (change in that character and possibly others). For DiffText, now ensure _all_ characters in the line (except indent) are actually different (so this is a stricter rule).